post index styling
