+++ src/components/PostIndex/PostList.scss
@@ -1,8 +1,60 @@
+.search-filter-new-post {
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  align-items: center;
+
+  .search-filter {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    align-items: center;
+
+    div.rmsc.multi-select .dropdown-container {
+      display: flex;
+      height: 56px;
+      border-color: rgba(0, 0, 0, 0.263);
+      justify-content: space-between;
+    
+        .dropdown-heading-value {
+          padding-left: 0.3em;
+        }
+
+    }
+  }
+
+  .new-post-container {
+    justify-self: right;
+
+    #new-post-btn {
+      margin-left: 0.5em;
+      height: 54px;
+      justify-self: right;
+    }
+
+    #new-post-link {
+      text-decoration: none;
+      color: white;
+    }
+   }
+
+}
+
+div.rmsc.multi-select {
+  width: 250px;
+  height: 56px;
+
+  .dropdown-container {
+    height: 55px;
+    display: flex;
+    align-items: center;
+  }
+}
+
 .post-list {
-  display: grid;
-  grid-template-columns: repeat(4, 1fr);
-  grid-auto-rows: auto;
-  grid-gap: 1rem;
-  // flex-direction: row;
-  // flex-wrap: wrap;
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  justify-content: center;
+  width: 100%;
 }
