+++ src/components/PostIndex/PostListItem.scss
@@ -3,7 +3,19 @@
   margin: 1em;
 
   .card-component {
-    width: 400;
+    width: 350px;
+
+    .css-185gdzj-MuiCardHeader-root {
+      padding-top: 0;
+    }
+
+    .MuiCardHeader-title {
+      font-size: 18px;
+    }
+
+    .MuiCardHeader-subheader {
+      font-size: 13px;
+    }
   }
 
   .heart-icon {
@@ -17,6 +29,12 @@
     align-items: center;
     margin-left: 0.5em;
     margin-right: 0.5em;
+
+    .likes-comments {
+      display: flex;
+      flex-direction: row;
+      align-items: center;
+    }
 
     span {
       display: flex;
